requirements.txt
django~=4.0
base58~=2.1.1
requests~=2.26.0
djangorestframework~=3.14.0
solana~=0.18.1
django-filter~=22.1
pillow~=9.2.0
unidecode~=1.3.6
python-dotenv~=0.21.0
django-nested-admin~=3.2.4